using a isWindows constant, rollback require in the example folder as it should just work
Move the install to a setup file and check for that and in case execute it
