+++ postinstall.js
@@ -10,40 +10,18 @@ const path = require('path');
 const fs = require('fs');
 
 let notNPM = true;
-
+const setup_file = path.resolve(__dirname, './setup.js');
 try {
-  fs.accessSync(path.resolve(__dirname, './website/package.json'));
+  fs.accessSync(setup_file);
 } catch (e) {
   notNPM = false;
 }
 
-const spawnSync = require('child_process').spawnSync;
-
-const execute = (cmd) => {
-  const options = {
-    stdio: 'inherit',
-    cwd: path.resolve(__dirname, cmd[0]),
-  };
-  spawnSync(cmd[1], cmd[2].split(' '), options);
-};
+const execute = require('./execute');
 
 if (
-  notNPM &&
-  process.env.NODE_ENV !== 'production'
+  notNPM ||
+  process.env.NODE_ENV === 'production'
 ) {
-  console.log(`Setting up Jest's development environment...`);
-  var npm = (process.platform === 'win32' ? 'npm.cmd' : 'npm');
-  var lerna = (process.platform === 'win32' ? 'lerna.cmd' : 'lerna');
-  const cmds = [
-    ['.', path.resolve(__dirname, './node_modules/.bin/' + lerna), 'bootstrap'],
-    ['packages/jest-jasmine1', npm, 'link'],
-    ['packages/jest-jasmine2', npm, 'link'],
-    ['packages/jest-mock', npm, 'link'],
-    ['packages/jest-util', npm, 'link'],
-    ['.', npm, 'link jest-jasmine1'],
-    ['.', npm, 'link jest-jasmine2'],
-    ['.', npm, 'link jest-mock'],
-    ['.', npm, 'link jest-util'],
-  ];
-  cmds.forEach(execute);
+  execute('.', 'node', setup_file);
 }
