+++ run_tests.js
@@ -19,6 +19,7 @@ const mkdirp = require('mkdirp');
 const PACKAGES_DIR = './packages';
 const EXAMPLES_DIR = './examples';
 const rimraf = require('rimraf');
+const isWindows = process.platform === 'win32';
 
 const packages = fs.readdirSync(PACKAGES_DIR)
   .map(file => path.resolve(PACKAGES_DIR, file))
@@ -48,29 +49,27 @@ packages.forEach(cwd => {
 examples.forEach(cwd => {
   console.log(chalk.bold(chalk.cyan('Testing example: ') + cwd));
 
-  runCommands([
-    'npm update',
-  ], cwd);
+  runCommands('npm update', cwd);
   rimraf.sync(path.resolve(cwd, './node_modules/jest-cli'));
   mkdirp.sync(path.resolve(cwd, './node_modules/jest-cli'));
   mkdirp.sync(path.resolve(cwd, './node_modules/.bin'));
 
-  const currentJest = path.resolve(__dirname, './bin/jest.js');
   // Using `npm link jest-cli` can create problems with module resolution,
   // so instead of this we'll create an `index.js` file that will export the
   // local `jest-cli` package.
   fs.writeFileSync(
     path.resolve(cwd, './node_modules/jest-cli/index.js'),
-    `module.exports = require('${__dirname.replace(/\\/g,'\\\\')}');\n`, // link to the local jest
+    `module.exports = require('../../../../');\n`, // link to the local jest
     'utf8'
   );
 
   // overwrite the jest link and point it to the local jest-cli
-  if (process.platform !== 'win32') {
-    runCommands(['ln -sf ../../bin/jest.js ./node_modules/.bin/jest'], cwd);
+  if (!isWindows) {
+    runCommands('ln -sf ../../bin/jest.js ./node_modules/.bin/jest', cwd);
   } else {
+    const currentJest = path.resolve(__dirname, './bin/jest.js');
     fs.writeFile(
-      '.\node_modules\.bin\jest.cmd',
+      `.\node_modules\.bin\jest.cmd`,
       `@IF EXIST "%~dp0\node.exe" (
         "%~dp0\node.exe"  "${currentJest}" %*
       ) ELSE (
@@ -86,6 +85,5 @@ examples.forEach(cwd => {
     );
 
   }
-  runCommands(['npm test'], cwd);
-
+  runCommands('npm test', cwd);
 });
